+++ cell.js
@@ -1,0 +1,5 @@
+export default class Cell{
+    constructor(value){
+        this.value=value;
+    }
+}
