+++ index.html
@@ -5,9 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>GetSchuwifty</title>
-    <script src="board.js"></script>
+    <script type="module" src="bootstrapper.js"></script>
 </head>
 <body>
-    <table id="board"></table>
+    <div id="game"></table>
 </body>
 </html>
